requirements.txt
streamlit
requests
beautifulsoup4
selenium
pillow
matplotlib
wordcloud
